requirements_dev.txt
asttokens==2.2.1
backcall==0.2.0
bpy==3.5.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
controlnet-aux==0.0.3
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
einops==0.6.1
executing==1.2.0
filelock==3.11.0
Flask==2.2.3
fonttools==4.39.3
gunicorn==20.1.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
importlib-metadata==6.5.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
scikit-image==0.20.0
scipy==1.10.1
segment-anything==1.0
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tifffile==2023.4.12
timm==0.6.13
tokenizers==0.13.3
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
zstandard==0.21.0
